attempt to fix integer size for 32bit

+++ src/types.jl
@@ -1,7 +1,7 @@
 # types for PRIMME.jl
 
 # FIXME: this should be exported from JLL
-const PRIMME_INT = Int
+const PRIMME_INT = Int64
 
 """
     PRIMMEException
